third_party/WebKit/LayoutTests/inspector/sources/debugger-ui/debugger-expand-scope.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script>

function makeClosure(n)
{
    var makeClosureLocalVar = "local." + n;
    return function innerFunction(x)
    {
        var innerFunctionLocalVar = x + 2;
        var negInf = -Infinity;
        var negZero = 1 / negInf;
        try {
            throw new Error("An exception");
        } catch (e) {
            e.toString();
            debugger;
        }
        return n + makeClosureLocalVar + x + innerFunctionLocalVar;
    }
}

function testFunction()
{
    var f = makeClosure("TextParam");
    f(2010);
}

function test()
{
    SourcesTestRunner.startDebuggerTest(onTestStarted);

    function onTestStarted()
    {
        TestRunner.addSniffer(Sources.ScopeChainSidebarPane.prototype, "_sidebarPaneUpdatedForTest", onSidebarRendered, true);
        SourcesTestRunner.runTestFunctionAndWaitUntilPaused(() => {});
    }

    function onSidebarRendered()
    {
        SourcesTestRunner.expandScopeVariablesSidebarPane(onScopeVariablesExpanded);
    }

    function onScopeVariablesExpanded()
    {
        TestRunner.addResult("");
        SourcesTestRunner.dumpScopeVariablesSidebarPane();
        SourcesTestRunner.completeDebuggerTest();
    }
}

</script>
</head>

<body onload="runTest()">
<input type='button' onclick='testFunction()' value='Test'/>
<p>
Test that sections representing scopes of the current call frame are expandable
and contain correct data.
</p>
</body>
</html>
