third_party/WebKit/LayoutTests/inspector/sources/debugger-ui/debugger-expand-scope-expected.txt


Test that sections representing scopes of the current call frame are expandable and contain correct data.

Set timer for test function.
Script execution paused.

Scope variables sidebar pane:
Catch
    e: Error: An exception
    at innerFunction (...)
    at testFunction (...) {stack: "Error: An exception↵    at innerFunction (...)", message: "An exception"}
Local
    innerFunctionLocalVar: 2012
    negInf: -Infinity
    negZero: -0
    this: Window
    x: 2010
Closure (makeClosure)
    makeClosureLocalVar: "local.TextParam"
    n: "TextParam"
WindowGlobal
    <section collapsed>
Script execution resumed.
